app/requirements.txt
aiohttp==3.1.0
aioredis==1.2.0
asgiref==3.1.4
asn1crypto==0.24.0
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
base58==1.0.3
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
Django==2.1.2
djangorestframework==3.8.2
drf-extensions==0.4.0
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
idna-ssl==1.1.0
immutables==0.9
importlib-metadata==0.18
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==7.2.0
msgpack==0.6.1
multidict==4.5.2
oauthlib==3.0.2
outcome==1.0.0
packaging==19.0
pathlib2==2.3.4
pluggy==0.12.0
psycopg2-binary==2.8.2
py==1.8.0
pycparser==2.19
pycurl==7.43.0
pygobject==3.20.0
PyHamcrest==1.9.0
pyparsing==2.4.0
pytest==5.0.1
pytest-asyncio==0.10.0
pytest-django==3.5.1
pytest-tornasync==0.6.0.post2
pytest-trio==0.5.2
python-apt==1.1.0b1+ubuntu0.16.4.5
python-dateutil==2.8.0
python-memcached==1.59
python3-indy==1.10.0
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.2.0
sentry-asgi==0.1.5
sentry-sdk==0.7.3
six==1.12.0
sniffio==1.1.0
sortedcontainers==2.1.0
teamcity-messages==1.21
tornado==6.0.3
trio==0.11.0
Twisted==19.2.1
txaio==18.8.1
typing-extensions==3.7.4
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
whitenoise==4.1.1
yarl==1.3.0
zipp==0.5.2
zope.interface==4.6.0
